chore: create oscd-layout addon... - extracts all rendering UI logic including former Hosting mixin and Plugging mixin to the layout addon - creates new events to connect layout addon to plugging loading functionality
